# Conflicts:
#	Assets/Turret/Prefabs/MiniGunTurret/MinigunTurret1.prefab

+++ Assets/Scripts/Item/Brricade.cs
@@ -32,9 +32,9 @@ public class Brricade : MonoBehaviour
         obstacle.enabled = false;
     }
 
-    public void Hurt(float damage)
+    public void Hurt(int damage)
     {
-        HP -= (int)damage;
+        HP -= damage;
 
         if(HP <= 0)
         {
